+++ CHANGES.md
@@ -46,6 +46,7 @@
 - Fix incorrect date displayed for support plan [Taiga #11986](https://tree.taiga.io/project/penpot/issue/11986)
 - Fix can't import 'borderWidth' type token [#132](https://github.com/tokens-studio/penpot/issues/132)
 - Fix moving elements up or down while pressing alt [Taiga Issue #11992](https://tree.taiga.io/project/penpot/issue/11992)
+- Fix conflicting shortcuts (remove dec/inc line height and letter spacing) [Taiga #12102](https://tree.taiga.io/project/penpot/issue/12102)
 
 ## 2.9.0